requirements.txt
psycopg2-binary
colorama>=0.4.6
pandas>=1.3.0
python-dotenv>=1.0.0
oracledb>=1.3.0
paramiko>=3.0.0
requests>=2.25.0
debugpy==1.8.5